Improvements for Subset handling.
1) Subsets are now case-insensitive (That is: soma and or Soma refer to
the same subset). Usually geometries should only contain "soma" in all
lower case, but accidently people name it all upper case or mixed.

If a subset is NOT found, a SubsetNotFoundException will be thrown and
indicate the problem, i.e. the name of the missing subset will be
printed.

2) See NeuronSimulation1D.cs please - Path setup did not work for me to
retrieve the test.vrn file.

3) Updated test.vrn: Broken meshes have been a result of not packacking
the triangulated mesh with mapping information but packaging the
quadrilateral mesh with mapping information. The packaging to .vrn (with
the triangulated mesh and associated mapping information) is fixed in
teh VR pipeline script for grid generation (See stephanmg/vr-grids repo)