Multi-Synapse
Added a complete new model for the synapse. Also restructured the framework of the code so it is more accessible to access the synaptic functions and easier to build off of. Also made it able to access the 1D nodes so we can start to apply voltages through the synapse.