Added preprocessor directives to the CellPreviewer.cs so that macOS users can see the cell window at runtime
There seems to be an inconsistency on what Path.DirectorySeparatorChar returns. Windows machines use the backslash to separate directories in file paths. While Unix-based machines use the forward slash to do that. In this case the back slash cause the mac machine not able to find the correct path to the cell geometries. Adding preprocessor directives to the file to enforce the correct slash is used on macOS fixed the issue.